Renaming module to be a bit more clean
