+++ cmd/graphbuild/dir_domain_emitter.go
@@ -1,4 +1,4 @@
-package graph
+package graphbuild
 
 import (
 	"compress/gzip"
