+++ cmd/root.go
@@ -3,7 +3,7 @@ package cmd
 import (
 	"os"
 
-	"github.com/bmeg/sifter/cmd/graph"
+	"github.com/bmeg/sifter/cmd/graphbuild"
 	"github.com/bmeg/sifter/cmd/graphfix"
 	"github.com/bmeg/sifter/cmd/graphmanifest"
 	"github.com/bmeg/sifter/cmd/manifest"
@@ -22,7 +22,7 @@ var RootCmd = &cobra.Command{
 func init() {
 	RootCmd.AddCommand(run.Cmd)
 	RootCmd.AddCommand(manifest.Cmd)
-	RootCmd.AddCommand(graph.Cmd)
+	RootCmd.AddCommand(graphbuild.Cmd)
 	RootCmd.AddCommand(template.Cmd)
 	RootCmd.AddCommand(graphmanifest.Cmd)
 	RootCmd.AddCommand(graphfix.Cmd)
